Implemented note recording and saving function
--Added functionality to start/stop note recording
--Implemented simulated transcription for notes
--Added modal for creating and saving notes
--Integrated basic UI updates for note-taking feature

Note: For speech to text react native requires bridging in its native OS for mobile and it was extremely tricky to troubleshoot, so I tried to fall back on a Google API for speech to text but it costs money and is not a long term solution. Eventually, we will integrate a way to use react native for speech to text because its library doesn't work properly with expo go or android emulator.